refactor: 패키지 수정
- StringRepeater를 repeater 하위 패키지에 두어 일관적인 구조를 가지도록 함
- util 패키지 하위의 클래스들이 유틸리티가 아닌 각 객체이므로 패키지명을 support로 수정
