+++ src/main/java/racingcar/Application.java
@@ -4,12 +4,12 @@ import camp.nextstep.edu.missionutils.Console;
 import racingcar.controller.RacingCarController;
 import racingcar.strategy.MovingStrategy;
 import racingcar.strategy.RacingCarMovingStrategy;
-import racingcar.util.StringRepeater;
-import racingcar.util.comparable.IntegerComparable;
-import racingcar.util.comparable.NumberComparable;
-import racingcar.util.random.RandomIntegerGenerator;
-import racingcar.util.random.RandomNumberGenerator;
-import racingcar.util.splitter.Splitter;
+import racingcar.support.repeater.StringRepeater;
+import racingcar.support.comparable.IntegerComparable;
+import racingcar.support.comparable.NumberComparable;
+import racingcar.support.random.RandomIntegerGenerator;
+import racingcar.support.random.RandomNumberGenerator;
+import racingcar.support.splitter.Splitter;
 import racingcar.view.ConsoleInputView;
 import racingcar.view.ConsoleOutputView;
 import racingcar.view.InputView;
