+++ src/main/java/racingcar/support/comparable/NumberComparable.java
@@ -1,4 +1,4 @@
-package racingcar.util.comparable;
+package racingcar.support.comparable;
 
 @FunctionalInterface
 public interface NumberComparable<T extends Number> {
