+++ src/main/java/racingcar/support/random/RandomNumberGenerator.java
@@ -1,4 +1,4 @@
-package racingcar.util.random;
+package racingcar.support.random;
 
 @FunctionalInterface
 public interface RandomNumberGenerator<T extends Number> {
