+++ src/test/java/racingcar/support/splitter/SplitterTest.java
@@ -1,4 +1,4 @@
-package racingcar.util.splitter;
+package racingcar.support.splitter;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
